refactor(editor): engine residual hooks 분리
