+++ front/src/components/editor/blockEditorEngineDocumentModel.ts
@@ -1,0 +1,107 @@
+import type { Editor as TiptapEditor } from "@tiptap/core"
+import type { KeyboardEvent as ReactKeyboardEvent } from "react"
+import type { BlockEditorDoc } from "./serialization"
+
+export const blockHasVisibleContent = (node?: BlockEditorDoc | null): boolean => {
+  if (!node) return false
+
+  if (node.type === "text") {
+    return Boolean((node as { text?: string }).text?.trim().length)
+  }
+
+  if (
+    node.type === "resizableImage" ||
+    node.type === "calloutBlock" ||
+    node.type === "taskList" ||
+    node.type === "bookmarkBlock" ||
+    node.type === "embedBlock" ||
+    node.type === "fileBlock" ||
+    node.type === "formulaBlock" ||
+    node.type === "toggleBlock" ||
+    node.type === "mermaidBlock" ||
+    node.type === "rawMarkdownBlock" ||
+    node.type === "table" ||
+    node.type === "horizontalRule"
+  ) {
+    return true
+  }
+
+  return Array.isArray(node.content) && node.content.some((child) => blockHasVisibleContent(child as BlockEditorDoc))
+}
+
+export const normalizeTableColorInputValue = (value: unknown) => {
+  const normalized = String(value || "").trim()
+  return /^#[0-9a-f]{6}$/i.test(normalized) ? normalized : "#dbeafe"
+}
+
+export const isPrimaryModifierPressed = (event: ReactKeyboardEvent | globalThis.KeyboardEvent) =>
+  event.metaKey || event.ctrlKey
+
+export const getActiveSlashRangeFromEditor = (editor: TiptapEditor): { from: number; to: number } | null => {
+  const selection = editor.state.selection
+  if (!selection.empty) return null
+
+  const parent = selection.$from.parent
+  if (parent.type.name !== "paragraph") return null
+
+  const textBeforeCursor = parent.textContent.slice(0, selection.$from.parentOffset)
+  const match = /(^|[\s\u00A0])\/([^\n]*)$/.exec(textBeforeCursor)
+  if (!match) return null
+
+  const slashOffset = (match.index ?? 0) + match[1].length
+  return {
+    from: selection.$from.start() + slashOffset,
+    to: selection.from,
+  }
+}
+
+export const focusElementWithoutScroll = (element: HTMLElement | null) => {
+  if (!element) return
+  if (typeof window === "undefined") {
+    element.focus()
+    return
+  }
+
+  const previousScrollX = window.scrollX
+  const previousScrollY = window.scrollY
+  try {
+    element.focus({ preventScroll: true })
+  } catch {
+    element.focus()
+  }
+  if (window.scrollX !== previousScrollX || window.scrollY !== previousScrollY) {
+    window.scrollTo(previousScrollX, previousScrollY)
+  }
+}
+
+export const resolveDocPosSafe = (editor: TiptapEditor, pos: number) => {
+  if (!Number.isFinite(pos)) return null
+  const normalizedPos = Math.round(pos)
+  const maxPos = editor.state.doc.content.size
+  if (normalizedPos < 0 || normalizedPos > maxPos) return null
+  try {
+    return editor.state.doc.resolve(normalizedPos)
+  } catch {
+    return null
+  }
+}
+
+export const downgradeDisabledFeatureNodes = (node: BlockEditorDoc, enableMermaidBlocks: boolean): BlockEditorDoc => {
+  if (!enableMermaidBlocks && node.type === "mermaidBlock") {
+    const source = String(node.attrs?.source || "").trim()
+    return {
+      type: "rawMarkdownBlock",
+      attrs: {
+        markdown: ["```mermaid", source, "```"].join("\n"),
+        reason: "unsupported-mermaid",
+      },
+    }
+  }
+
+  if (!node.content?.length) return node
+
+  return {
+    ...node,
+    content: node.content.map((child) => downgradeDisabledFeatureNodes(child as BlockEditorDoc, enableMermaidBlocks)),
+  }
+}
