+++ front/src/components/editor/useBlockEditorEngineSelectionBubbleEffects.ts
@@ -1,0 +1,296 @@
+import type { Editor as TiptapEditor } from "@tiptap/core"
+import { TextSelection } from "@tiptap/pm/state"
+import { useEffect } from "react"
+import type { Dispatch, MutableRefObject, RefObject, SetStateAction } from "react"
+import { preserveWindowScrollForEditorPointerFocus } from "./blockHandleLayoutModel"
+import { isTableSelectionActive } from "./tableStructureModel"
+import {
+  areFloatingBubbleStatesEqual,
+  hideFloatingBubbleState,
+  resolveFloatingBubbleStateFromCoords,
+  type FloatingBubbleState,
+} from "./useFloatingBubbleState"
+
+type SetState<T> = Dispatch<SetStateAction<T>>
+
+type UseBlockEditorEngineSelectionBubbleEffectsArgs = {
+  bubbleToolbarHoveredRef: RefObject<boolean>
+  cancelBubbleHide: () => void
+  clearWindowTextSelection: () => void
+  editor: TiptapEditor | null
+  editorRef: RefObject<TiptapEditor | null>
+  hasTableStructuralSelection: (editor: TiptapEditor) => boolean
+  hideTableQuickRailImmediately: () => void
+  isTableColumnRailResizeActive: () => boolean
+  mouseTextSelectionInProgressRef: MutableRefObject<boolean>
+  scheduleBubbleHide: () => void
+  scheduleTableQuickRailHide: () => void
+  setBubbleState: SetState<FloatingBubbleState>
+  syncBubbleOnMouseUpRef: MutableRefObject<boolean>
+  syncTableQuickRailFromElement: (element: Element, clientX?: number, clientY?: number) => void
+  tableMenuState: unknown
+  tryStartTableColumnResizeFromDomHandle: (target: EventTarget | null, pointerId: number, clientX: number) => boolean
+}
+
+export const useBlockEditorEngineSelectionBubbleEffects = ({
+  bubbleToolbarHoveredRef,
+  cancelBubbleHide,
+  clearWindowTextSelection,
+  editor,
+  editorRef,
+  hasTableStructuralSelection,
+  hideTableQuickRailImmediately,
+  isTableColumnRailResizeActive,
+  mouseTextSelectionInProgressRef,
+  scheduleBubbleHide,
+  scheduleTableQuickRailHide,
+  setBubbleState,
+  syncBubbleOnMouseUpRef,
+  syncTableQuickRailFromElement,
+  tableMenuState,
+  tryStartTableColumnResizeFromDomHandle,
+}: UseBlockEditorEngineSelectionBubbleEffectsArgs) => {
+  useEffect(() => {
+    const currentEditor = editorRef.current ?? editor
+    if (!currentEditor) return
+    let rafId: number | null = null
+
+    const syncBubble = () => {
+      const activeEditor = editorRef.current ?? currentEditor
+      if (!activeEditor) {
+        scheduleBubbleHide()
+        if (!tableMenuState) {
+          hideTableQuickRailImmediately()
+        }
+        return
+      }
+
+      let selection = activeEditor.state.selection
+      if (selection.empty && typeof window !== "undefined" && !isTableColumnRailResizeActive()) {
+        const domSelection = window.getSelection()
+        const range =
+          domSelection && domSelection.rangeCount > 0 ? domSelection.getRangeAt(0) : null
+        const commonAncestor =
+          range?.commonAncestorContainer instanceof Element
+            ? range.commonAncestorContainer
+            : range?.commonAncestorContainer?.parentElement ?? null
+
+        if (range && domSelection && !domSelection.isCollapsed && commonAncestor && activeEditor.view.dom.contains(commonAncestor)) {
+          const syncPmSelectionFromRange = (from: number, to: number) => {
+            if (!Number.isFinite(from) || !Number.isFinite(to) || from === to) return false
+            const nextSelection = TextSelection.create(
+              activeEditor.state.doc,
+              Math.min(from, to),
+              Math.max(from, to)
+            )
+            if (!nextSelection.eq(activeEditor.state.selection)) {
+              activeEditor.view.dispatch(activeEditor.state.tr.setSelection(nextSelection))
+              selection = activeEditor.state.selection
+            }
+            return true
+          }
+
+          let synced = false
+          try {
+            const from = activeEditor.view.posAtDOM(range.startContainer, range.startOffset)
+            const to = activeEditor.view.posAtDOM(range.endContainer, range.endOffset)
+            synced = syncPmSelectionFromRange(from, to)
+          } catch {
+            synced = false
+          }
+
+          if (!synced) {
+            const rangeRects = Array.from(range.getClientRects())
+            const startRect = rangeRects[0] ?? range.getBoundingClientRect()
+            const endRect = rangeRects[rangeRects.length - 1] ?? startRect
+            const startCoords = activeEditor.view.posAtCoords({
+              left: startRect.left + 1,
+              top: startRect.top + startRect.height / 2,
+            })
+            const endCoords = activeEditor.view.posAtCoords({
+              left: Math.max(endRect.left + 1, endRect.right - 1),
+              top: endRect.top + endRect.height / 2,
+            })
+            if (startCoords?.pos && endCoords?.pos) {
+              syncPmSelectionFromRange(startCoords.pos, endCoords.pos)
+            }
+          }
+        }
+      }
+
+      const isImageNodeSelected = activeEditor.isActive("resizableImage")
+      const isTableActive = isTableSelectionActive(activeEditor)
+      const isTableStructuralSelection = hasTableStructuralSelection(activeEditor)
+      const canShowTextToolbar =
+        !selection.empty &&
+        !isImageNodeSelected &&
+        !activeEditor.isActive("codeBlock") &&
+        !activeEditor.isActive("rawMarkdownBlock") &&
+        !isTableStructuralSelection
+
+      if (canShowTextToolbar && mouseTextSelectionInProgressRef.current) {
+        syncBubbleOnMouseUpRef.current = true
+        if (bubbleToolbarHoveredRef.current) return
+        setBubbleState((prev) =>
+          prev.visible && prev.mode === "text" ? hideFloatingBubbleState(prev) : prev
+        )
+        if (!tableMenuState) {
+          hideTableQuickRailImmediately()
+        }
+        return
+      }
+
+      if (!isImageNodeSelected && !canShowTextToolbar && !isTableActive) {
+        if (bubbleToolbarHoveredRef.current) return
+        scheduleBubbleHide()
+        if (!tableMenuState) {
+          scheduleTableQuickRailHide()
+        }
+        return
+      }
+
+      if (isTableActive && !canShowTextToolbar) {
+        cancelBubbleHide()
+        setBubbleState(hideFloatingBubbleState)
+        const anchorDom = activeEditor.view.domAtPos(selection.from).node
+        const anchorElement =
+          anchorDom instanceof Element ? anchorDom : anchorDom.parentElement
+        if (isTableStructuralSelection && anchorElement?.closest(".aq-table-shell, .tableWrapper, table")) {
+          syncTableQuickRailFromElement(anchorElement)
+          return
+        }
+        if (!tableMenuState) {
+          hideTableQuickRailImmediately()
+        }
+        return
+      }
+
+      cancelBubbleHide()
+      hideTableQuickRailImmediately()
+
+      const startCoords = activeEditor.view.coordsAtPos(selection.from)
+      const endCoords = activeEditor.view.coordsAtPos(isImageNodeSelected ? selection.from : selection.to)
+      const nextBubbleState = resolveFloatingBubbleStateFromCoords(
+        isImageNodeSelected ? "image" : "text",
+        startCoords,
+        endCoords
+      )
+      setBubbleState((prev) =>
+        areFloatingBubbleStatesEqual(prev, nextBubbleState) ? prev : nextBubbleState
+      )
+    }
+
+    const scheduleSyncBubble = () => {
+      if (typeof window === "undefined") {
+        syncBubble()
+        return
+      }
+      if (rafId !== null) return
+      rafId = window.requestAnimationFrame(() => {
+        rafId = null
+        syncBubble()
+      })
+    }
+
+    const handleDocumentSelectionChange = () => {
+      const activeEditor = editorRef.current ?? currentEditor
+      if (!activeEditor) return
+      if (isTableColumnRailResizeActive()) {
+        clearWindowTextSelection()
+        return
+      }
+      const selection = window.getSelection()
+      const anchorNode = selection?.anchorNode ?? null
+      const anchorElement =
+        anchorNode instanceof Element ? anchorNode : anchorNode?.parentElement ?? null
+      if (anchorElement && !activeEditor.view.dom.contains(anchorElement)) return
+      scheduleSyncBubble()
+    }
+
+    const handleEditorPointerDownCapture = (event: PointerEvent) => {
+      if (event.pointerType !== "mouse" || event.button !== 0) return
+      const activeEditor = editorRef.current ?? currentEditor
+      if (!activeEditor) return
+      if (!(event.target instanceof Node) || !activeEditor.view.dom.contains(event.target)) return
+      preserveWindowScrollForEditorPointerFocus(event.target, isTableSelectionActive(activeEditor))
+      if (tryStartTableColumnResizeFromDomHandle(event.target, event.pointerId, event.clientX)) {
+        event.preventDefault()
+        event.stopPropagation()
+        event.stopImmediatePropagation?.()
+        mouseTextSelectionInProgressRef.current = false
+        syncBubbleOnMouseUpRef.current = false
+        return
+      }
+      mouseTextSelectionInProgressRef.current = true
+      syncBubbleOnMouseUpRef.current = false
+      if (bubbleToolbarHoveredRef.current) return
+      setBubbleState((prev) =>
+        prev.visible && prev.mode === "text" ? hideFloatingBubbleState(prev) : prev
+      )
+    }
+
+    const handleEditorMouseDownCapture = (event: MouseEvent) => {
+      const activeEditor = editorRef.current ?? currentEditor
+      if (!activeEditor) return
+      if (!(event.target instanceof Node) || !activeEditor.view.dom.contains(event.target)) return
+      if (!(event.target instanceof Element) || !event.target.closest(".column-resize-handle")) return
+      event.preventDefault()
+      event.stopPropagation()
+      event.stopImmediatePropagation?.()
+    }
+
+    const handleWindowPointerUp = (event: PointerEvent) => {
+      if (!mouseTextSelectionInProgressRef.current) return
+      if (event.pointerType && event.pointerType !== "mouse") return
+      mouseTextSelectionInProgressRef.current = false
+      if (!syncBubbleOnMouseUpRef.current) return
+      syncBubbleOnMouseUpRef.current = false
+      scheduleSyncBubble()
+    }
+
+    scheduleSyncBubble()
+    currentEditor.on("selectionUpdate", scheduleSyncBubble)
+    currentEditor.on("focus", scheduleSyncBubble)
+    document.addEventListener("selectionchange", handleDocumentSelectionChange)
+    document.addEventListener("pointerdown", handleEditorPointerDownCapture, true)
+    document.addEventListener("mousedown", handleEditorMouseDownCapture, true)
+    window.addEventListener("scroll", scheduleSyncBubble, { capture: true, passive: true })
+    window.addEventListener("resize", scheduleSyncBubble, { passive: true })
+    window.addEventListener("pointerup", handleWindowPointerUp, true)
+    window.addEventListener("pointercancel", handleWindowPointerUp, true)
+    return () => {
+      currentEditor.off("selectionUpdate", scheduleSyncBubble)
+      currentEditor.off("focus", scheduleSyncBubble)
+      document.removeEventListener("selectionchange", handleDocumentSelectionChange)
+      document.removeEventListener("pointerdown", handleEditorPointerDownCapture, true)
+      document.removeEventListener("mousedown", handleEditorMouseDownCapture, true)
+      window.removeEventListener("scroll", scheduleSyncBubble, true)
+      window.removeEventListener("resize", scheduleSyncBubble)
+      window.removeEventListener("pointerup", handleWindowPointerUp, true)
+      window.removeEventListener("pointercancel", handleWindowPointerUp, true)
+      if (rafId !== null && typeof window !== "undefined") {
+        window.cancelAnimationFrame(rafId)
+      }
+      mouseTextSelectionInProgressRef.current = false
+      syncBubbleOnMouseUpRef.current = false
+      cancelBubbleHide()
+    }
+  }, [
+    bubbleToolbarHoveredRef,
+    cancelBubbleHide,
+    clearWindowTextSelection,
+    editor,
+    editorRef,
+    hasTableStructuralSelection,
+    hideTableQuickRailImmediately,
+    isTableColumnRailResizeActive,
+    mouseTextSelectionInProgressRef,
+    scheduleBubbleHide,
+    scheduleTableQuickRailHide,
+    setBubbleState,
+    syncBubbleOnMouseUpRef,
+    syncTableQuickRailFromElement,
+    tableMenuState,
+    tryStartTableColumnResizeFromDomHandle,
+  ])
+}
